requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.3.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
openpyxl==3.1.2
reportlab==4.1.0
pandas==2.2.0
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-dateutil==2.8.2
Pillow==10.2.0